pyproject.toml
[project]
name = "llmx"
version = "0.6.0"
description = "Unified CLI and API for LLM providers via native SDKs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.50.0",
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "google-genai>=1.0.0",
    "pillow>=10.0.0",
]

[project.scripts]
llmx = "llmx.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = []
